feat(core): introduce `Render` component for flexible element composition
This commit introduces the `Render` component, a powerful internal utility designed to enhance the flexibility and composability of **Rilix UI** components.
`Render` allows consumers to customize the underlying HTML element of a component or seamlessly integrate its functionality with their own custom elements. This promotes a more modular and extensible approach to **Rilix UI** component design.