feat(core)!: remove 'hiddenUntilFound' prop from `Collapsible.Panel` component
The 'hiddenUntilFound' prop has been removed from the `Collapsible.Panel` component, as well as its related logic and context.
This prop, which previously controlled the 'hidden="until-found"' attribute, is no longer supported.

This is a breaking change, as any usage of the 'hiddenUntilFound' prop on `Collapsible.Panel` will now result in a TypeScript error.
Users should remove this prop from their component usage.