docs(core): improve clarity in `PortalContextValue` JSDoc
Refactors the JSDoc for the `PortalContextValue` interface to replace "imperatively" with "programmatically". This change provides a more precise and neutral description of how the portal's state can be controlled with code, aligning better with the declarative nature of frameworks.

The change was applied to the description of both the `PortalContextValue`
interface and the `setOpen$` function within it.