refactor(core): correct `disabled` type in `state` object of `Link.Root`'s `render$` prop
This change updates the type of the `disabled` property within the `state` object, which is passed to the `render$` prop of the `Link.Root` component.
The type is changed from `ReadonlySignal<boolean | undefined>` to `ReadonlySignal<boolean>`, reflecting the internal logic that now guarantees a boolean value, thereby preventing it from ever being `undefined`.