feat(core)!: remove unused `state` property from `render$` prop in `AspectRatio.Content` component
Removed the `aspect` property from the `state` object passed to the `render$` prop in `AspectRatio.Content`.
This simplifies the component's API and code by removing a property that was not being used.

This is a BREAKING CHANGE as the `aspect` property is no longer available in the `state` object passed to the `render$` prop.