feat(core): set initial focus on the first enabled alert dialog cancel button
The `AlertDialog.Content` component has been updated to automatically set focus to the first enabled cancel button when the dialog is opened.
This change ensures a better user experience and improves accessibility by directing the user's focus to a relevant, interactive element.

To achieve this, a new selector, `[data-rilix-ui-alert-dialog-cancel]:not([data-disabled])`, was introduced.
It specifically targets cancel buttons (`[data-rilix-ui-alert-dialog-cancel]`) and filters out any that are in a disabled state (`:not([data-disabled])`).
This prevents the focus from being set on non-interactive elements, which was an issue in the previous implementation.